+++ CHANGELOG.md
@@ -9,6 +9,8 @@ For older change log history see the [historic changelog](HISTORIC_CHANGELOG.md)
 
 ### Added
 
+- Added SccmIniFile resource
+
 ### Changed
 
 ### Removed